apm-data: remove '<data_stream.type>-apm@custom'
Stop invoking the '<data_stream.type>-apm@custom' ingest
pipeline (if it exists). This was done to align with what
Fleet was doing, which turned out to be flawed due to
naming collisions. e.g. this would mean that the data stream
`traces-apm.rum-default` would invoke `traces-apm@custom`,
which would also match the pipeline intended to be invoked
only for the data stream `traces-apm-default`.

The Fleet solution was to rename the convention to
`<data_stream.type>-<package>.integration@custom`. This does
not make sense for the apm-data plugin, and we see no use
case for this anyway. The behaviour was added in 8.12.1 and
has not been documented, so this is non-breaking.

See also
https://github.com/elastic/kibana/issues/175254#issuecomment-1909666258